+++ app/src/main/java/hr/dtakac/prognoza/ui/forecast/ComingItem.kt
@@ -1,0 +1,174 @@
+package hr.dtakac.prognoza.ui.forecast
+
+import android.content.Context
+import androidx.compose.animation.*
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.updateTransition
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.material.ripple.rememberRipple
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.ProvideTextStyle
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.platform.LocalFontFamilyResolver
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import coil.compose.rememberAsyncImagePainter
+import hr.dtakac.prognoza.R
+import hr.dtakac.prognoza.presentation.asString
+import hr.dtakac.prognoza.presentation.forecast.DayUi
+import hr.dtakac.prognoza.ui.theme.PrognozaTheme
+
+@Composable
+fun rememberComingItemDimensions(
+    days: List<DayUi>,
+    context: Context = LocalContext.current,
+    density: Density = LocalDensity.current,
+    fontFamilyResolver: FontFamily.Resolver = LocalFontFamilyResolver.current,
+    style: TextStyle = PrognozaTheme.typography.body
+): ComingItemDimensions = remember(days, context, density, fontFamilyResolver, style) {
+    ComingItemDimensions(
+        precipitationWidth = calculateMaxWidth(
+            texts = days.map { it.precipitation.asString(context) },
+            density = density,
+            fontFamilyResolver = fontFamilyResolver,
+            style = style
+        ),
+        lowHighTemperatureWidth = calculateMaxWidth(
+            texts = days.map { it.lowHighTemperature.asString(context) },
+            density = density,
+            fontFamilyResolver = fontFamilyResolver,
+            style = style
+        )
+    )
+}
+
+@OptIn(ExperimentalAnimationApi::class)
+@Composable
+fun ComingItem(
+    day: DayUi,
+    dimensions: ComingItemDimensions,
+    isExpanded: Boolean = false,
+    onClick: () -> Unit = {},
+    modifier: Modifier = Modifier
+) {
+    val expandTransition = updateTransition(label = "Expand coming item", targetState = isExpanded)
+    val chevronRotation by expandTransition.animateFloat(label = "Rotate chevron") {
+        if (it) 180f else 0f
+    }
+    Column(
+        modifier = Modifier
+            .clickable(
+                onClick = onClick,
+                indication = rememberRipple(bounded = true),
+                interactionSource = remember { MutableInteractionSource() }
+            )
+            .then(modifier)
+    ) {
+        ProvideTextStyle(PrognozaTheme.typography.body) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    modifier = Modifier.weight(1f),
+                    text = day.date.asString(),
+                    textAlign = TextAlign.Start,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+                Text(
+                    modifier = Modifier.width(dimensions.precipitationWidth),
+                    text = day.precipitation.asString(),
+                    textAlign = TextAlign.End,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    color = LocalContentColor.current.copy(alpha = PrognozaTheme.alpha.medium)
+                )
+                Spacer(modifier = Modifier.width(16.dp))
+                Text(
+                    modifier = Modifier.width(dimensions.lowHighTemperatureWidth),
+                    text = day.lowHighTemperature.asString(),
+                    textAlign = TextAlign.End,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+                Spacer(modifier = Modifier.width(4.dp))
+                Image(
+                    painter = painterResource(id = R.drawable.ic_expand_more),
+                    contentDescription = null,
+                    modifier = Modifier
+                        .size(24.dp)
+                        .graphicsLayer { rotationZ = chevronRotation },
+                    colorFilter = ColorFilter.tint(LocalContentColor.current)
+                )
+            }
+        }
+        ProvideTextStyle(PrognozaTheme.typography.bodySmall) {
+            expandTransition.AnimatedVisibility(
+                visible = { targetIsExpanded -> targetIsExpanded },
+                enter = fadeIn() + expandVertically(),
+                exit = fadeOut() + shrinkVertically()
+            ) {
+                LazyRow(
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    modifier = Modifier.padding(top = 16.dp)
+                ) {
+                    itemsIndexed(day.hours) { idx, hour ->
+                        Column(
+                            horizontalAlignment = Alignment.CenterHorizontally,
+                            modifier = when (idx) {
+                                0 -> Modifier.padding(end = 8.dp)
+                                day.hours.lastIndex -> Modifier.padding(start = 8.dp)
+                                else -> Modifier.padding(horizontal = 8.dp)
+                            }
+                        ) {
+                            Text(
+                                text = hour.temperature.asString(),
+                                maxLines = 1,
+                                overflow = TextOverflow.Ellipsis,
+                                modifier = Modifier.padding(bottom = 4.dp)
+                            )
+                            Image(
+                                painter = rememberAsyncImagePainter(model = hour.icon),
+                                contentDescription = null,
+                                modifier = Modifier.size(32.dp)
+                            )
+                            Text(
+                                text = hour.time.asString(),
+                                maxLines = 1,
+                                overflow = TextOverflow.Ellipsis,
+                                modifier = Modifier.padding(top = 4.dp),
+                                color = LocalContentColor.current
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+data class ComingItemDimensions(
+    val precipitationWidth: Dp,
+    val lowHighTemperatureWidth: Dp
+)
